More refactor and utilising "g" object
